Refactor
Model, Cell, View, ViewController 그룹 추가 및 파일, 코드 리팩터링
